References command and rplugin work around.
It seems that a function in remote plugin returns v:null the first time
it is called. To get around this it is called again if the result is
empty.
As an aside this will still work if the result is truly empty as it will
return empty twice.

Created the References command which uses fzf to select a reference and
returns the reference.